+++ test/Unit/Parsers/ProgrammaticVPAIDParserTest.ts
@@ -1,4 +1,5 @@
 import { AuctionResponse } from 'Ads/Models/AuctionResponse';
+import { AuctionPlacement } from 'Ads/Models/AuctionPlacement';
 import { Session } from 'Ads/Models/Session';
 import { assert } from 'chai';
 
@@ -15,7 +16,7 @@ import { VPAIDCampaign } from 'VPAID/Models/VPAIDCampaign';
 import { ProgrammaticVPAIDParser } from 'VPAID/Parsers/ProgrammaticVPAIDParser';
 
 describe('ProgrammaticVPAIDParser', () => {
-    const placements = ['TestPlacement'];
+    const placementId = 'TestPlacement';
     const mediaId = 'o2YMT0Cmps6xHiOwNMeCrH';
     const correlationId = '583dfda0d933a3630a53249c';
 
@@ -38,7 +39,8 @@ describe('ProgrammaticVPAIDParser', () => {
         describe('with proper XML payload', () => {
             let campaign: VPAIDCampaign;
             const parse = (data: any) => {
-                const response = new AuctionResponse(placements, data, mediaId, correlationId);
+                const auctionPlacement = new AuctionPlacement(placementId, mediaId);
+                const response = new AuctionResponse([auctionPlacement], data, mediaId, correlationId);
                 return parser.parse(nativeBridge, request, response, session).then((parsedCampaign) => {
                     campaign = <VPAIDCampaign>parsedCampaign;
                 });
@@ -57,7 +59,7 @@ describe('ProgrammaticVPAIDParser', () => {
 
                 assert.equal(campaign.getSession(), session, 'Session is not equal');
                 assert.equal(campaign.getMediaId(), mediaId, 'MediaID is not equal');
-                assert.equal(campaign.getVPAID().getScriptUrl(), 'https://fake-ads-backend.applifier.info/get_file/js/vpaid_sample.js', 'Script URL is not equal');
+                assert.equal(campaign.getVPAID().getScriptUrl(), 'https://fake-ads-backend.unityads.unity3d.com/get_file/js/vpaid_sample.js', 'Script URL is not equal');
             });
         });
     });